# openapi.json
{
  "components": {
    "schemas": {
      "AuthStatusOutputBody": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/AuthStatusOutputBody.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "available": {
            "description": "Whether Application Default Credentials are available.",
            "type": "boolean"
          },
          "message": {
            "description": "Sanitized credential status message.",
            "type": "string"
          }
        },
        "required": [
          "available",
          "message"
        ],
        "type": "object"
      },
...
      "ProfileResponse": {
        "additionalProperties": false,
        "properties": {
          "$schema": {
            "description": "A URL to the JSON Schema for this object.",
            "examples": [
              "https://example.com/ProfileResponse.json"
            ],
            "format": "uri",
            "readOnly": true,
            "type": "string"
          },
          "id": {
            "description": "Stable local connection profile ID.",
            "type": "string"
          },
          "name": {
            "description": "User-visible connection name.",
            "type": "string"
          },
          "projectId": {
            "description": "GCP project ID.",
            "type": "string"
          },
          "provider": {
            "description": "Cloud provider.",
            "enum": [
              "gcp"
            ],
            "type": "string"
          },
          "status": {
            "description": "Current local credential status.",
            "enum": [
              "ready",
              "needs-auth"
            ],
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "projectId",
          "provider",
          "status"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "session": {
        "description": "Local pairing session",
        "in": "cookie",
        "name": "logleopard_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "LogLeopard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/status": {
      "get": {
        "description": "Report sanitized GCP Application Default Credential availability.",
        "operationId": "auth-status",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthStatusOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get credential status",
        "tags": [
          "LogLeopard"
        ]
      }
    },
    "/api/v1/profiles": {
      "get": {
        "description": "List local GCP connection profiles without credentials.",
        "operationId": "list-profiles",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileResponse"
                  },
                  "type": [
                    "array",
                    "null"
                  ]
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List connection profiles",
        "tags": [
          "LogLeopard"
        ]
      },
      "post": {
        "description": "Create a local profile containing a display name and GCP project ID.",
        "operationId": "create-profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Create connection profile",
        "tags": [
          "LogLeopard"
        ]
      }
    },
    "/api/v1/profiles/{id}": {
      "delete": {
        "description": "Delete a local connection profile; cloud resources are not changed.",
        "operationId": "delete-profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Delete connection profile",
        "tags": [
          "LogLeopard"
        ]
      },
      "put": {
        "description": "Replace the name and GCP project ID of a local connection profile.",
        "operationId": "update-profile",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Update connection profile",
        "tags": [
          "LogLeopard"
        ]
      }
    },
    "/api/v1/query": {
      "post": {
        "description": "Query a bounded time window of Cloud Run revision logs and return normalized entries plus an opaque cursor.",
        "operationId": "query-logs",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Query Cloud Run logs",
        "tags": [
          "LogLeopard"
        ]
      }
    },
    "/api/v1/request-context": {
      "post": {
        "description": "Find up to 200 Cloud Run revision entries across all services in the profile project, within 15 minutes before or after the selected event, by exact known request ID or trace ID locations.",
        "operationId": "request-context",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestContextInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestContextOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Get request context",
        "tags": [
          "LogLeopard"
        ]
      }
    },
    "/api/v1/session/logout": {
      "post": {
        "description": "Invalidate the current local session and expire its cookie.",
        "operationId": "logout",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CookieOutputBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Log out",
        "tags": [
          "LogLeopard"
        ]
      }
    },
    "/api/v1/session/pair": {
      "post": {
        "description": "Exchange the single-use startup token for a local HttpOnly session cookie.",
        "operationId": "pair",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CookieOutputBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorModel"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Pair browser session",
        "tags": [
          "LogLeopard"
        ]
      }
    },
...
  }
}